+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - MSSQL - Prevent potential bad database connection using upgrading an old version of `tedious` dependency.
 - Command Generate - Catch correctly authentication error on mongodb.
 - Command Generate - Catch correctly empty collections on mongodb.
+- Command Generate - Fix connection to database with instance name.
 
 ## RELEASE 2.5.0 - 2019-10-15
 ### Added